+++ CHANGELOG.rst
@@ -2,4 +2,6 @@
 0.5.2
 ######
 
+- Add support for PostgreSQL JSONField, HStoreField, and ArrayField
+
 - Add Python 3+ compatibility